move the module sync methods (except era) to AccessSyncService
