+++ api/src/main/java/org/pmiops/workbench/db/dao/UserServiceImpl.java
@@ -5,10 +5,8 @@ import com.google.common.collect.ImmutableList;
 import java.sql.Timestamp;
 import java.time.Clock;
 import java.time.Instant;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Random;
 import java.util.Set;
@@ -21,14 +19,11 @@ import javax.annotation.Nonnull;
 import javax.inject.Provider;
 import javax.mail.MessagingException;
 import org.hibernate.exception.GenericJDBCException;
-import org.pmiops.workbench.access.AccessModuleNameMapper;
 import org.pmiops.workbench.access.AccessModuleService;
 import org.pmiops.workbench.access.AccessSyncService;
 import org.pmiops.workbench.access.AccessTierService;
 import org.pmiops.workbench.actionaudit.Agent;
 import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
-import org.pmiops.workbench.compliance.ComplianceService;
-import org.pmiops.workbench.compliance.ComplianceService.BadgeName;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.model.DbAccessModule.DbAccessModuleName;
 import org.pmiops.workbench.db.model.DbAddress;
@@ -54,11 +49,9 @@ import org.pmiops.workbench.monitoring.GaugeDataCollector;
 import org.pmiops.workbench.monitoring.MeasurementBundle;
 import org.pmiops.workbench.monitoring.labels.MetricLabel;
 import org.pmiops.workbench.monitoring.views.GaugeMetric;
-import org.pmiops.workbench.moodle.model.BadgeDetailsV2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpStatus;
 import org.springframework.orm.ObjectOptimisticLockingFailureException;
 import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Service;
@@ -88,9 +81,7 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
   private final VerifiedInstitutionalAffiliationDao verifiedInstitutionalAffiliationDao;
 
   private final AccessTierService accessTierService;
-  private final AccessModuleNameMapper accessModuleNameMapper;
   private final AccessModuleService accessModuleService;
-  private final ComplianceService complianceService;
   private final DirectoryService directoryService;
   private final FireCloudService fireCloudService;
   private final MailService mailService;
@@ -108,10 +99,8 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
       UserDao userDao,
       UserTermsOfServiceDao userTermsOfServiceDao,
       VerifiedInstitutionalAffiliationDao verifiedInstitutionalAffiliationDao,
-      AccessModuleNameMapper accessModuleNameMapper,
       AccessModuleService accessModuleService,
       FireCloudService fireCloudService,
-      ComplianceService complianceService,
       DirectoryService directoryService,
       AccessTierService accessTierService,
       MailService mailService,
@@ -124,10 +113,8 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
     this.userDao = userDao;
     this.userTermsOfServiceDao = userTermsOfServiceDao;
     this.verifiedInstitutionalAffiliationDao = verifiedInstitutionalAffiliationDao;
-    this.accessModuleNameMapper = accessModuleNameMapper;
     this.accessModuleService = accessModuleService;
     this.fireCloudService = fireCloudService;
-    this.complianceService = complianceService;
     this.directoryService = directoryService;
     this.accessTierService = accessTierService;
     this.mailService = mailService;
@@ -194,10 +181,6 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
         }
       }
     }
-  }
-
-  private boolean isServiceAccount(DbUser user) {
-    return configProvider.get().auth.serviceAccountApiUsers.contains(user.getUsername());
   }
 
   @Override
@@ -481,115 +464,6 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
   public Set<DbUser> findActiveUsersByUsernames(List<String> usernames) {
     return userDao.findUserByUsernameInAndDisabledFalse(usernames);
   }
-
-  /** Syncs the current user's training status from Moodle. */
-  @Override
-  public DbUser syncComplianceTrainingStatusV2()
-      throws org.pmiops.workbench.moodle.ApiException, NotFoundException {
-    DbUser user = userProvider.get();
-    return syncComplianceTrainingStatusV2(user, Agent.asUser(user));
-  }
-
-  /**
-   * Updates the given user's training status from Moodle.
-   *
-   * <p>We can fetch Moodle data for arbitrary users since we use an API key to access Moodle,
-   * rather than user-specific OAuth tokens.
-   *
-   * <p>Using the user's email, we can get their badges from Moodle's APIs. If the badges are marked
-   * valid, we store their completion dates in the database. If they are marked invalid, we clear
-   * the completion dates from the database as the user will need to complete a new training.
-   */
-  @Override
-  public DbUser syncComplianceTrainingStatusV2(DbUser dbUser, Agent agent)
-      throws org.pmiops.workbench.moodle.ApiException, NotFoundException {
-    // Skip sync for service account user rows.
-    if (isServiceAccount(dbUser)) {
-      return dbUser;
-    }
-
-    try {
-      Map<BadgeName, BadgeDetailsV2> userBadgesByName =
-          complianceService.getUserBadgesByBadgeName(dbUser.getUsername());
-
-      /**
-       * Determine the logical completion time for this user for the given compliance access module.
-       * Three logical outcomes are possible:
-       *
-       * <ul>
-       *   <li>Incomplete or invalid training badge: empty
-       *   <li>Badge has been issued for the first time, or has been reissued since we last marked
-       *       the training complete: now
-       *   <li>Else: existing completion time, i.e. no change
-       * </ul>
-       */
-      Function<BadgeName, Optional<Timestamp>> determineCompletionTime =
-          (badgeName) -> {
-            Optional<BadgeDetailsV2> badge =
-                Optional.ofNullable(userBadgesByName.get(badgeName))
-                    .filter(BadgeDetailsV2::getValid);
-
-            if (!badge.isPresent()) {
-              return Optional.empty();
-            }
-
-            if (badge.get().getLastissued() == null) {
-              log.warning(
-                  String.format(
-                      "badge %s is indicated as valid by Moodle, but is missing the lastissued "
-                          + "time, this is unexpected - treating this as an incomplete training",
-                      badgeName));
-              return Optional.empty();
-            }
-            Instant badgeTime = Instant.ofEpochSecond(badge.get().getLastissued());
-            Instant dbCompletionTime =
-                accessModuleService
-                    .getAccessModuleStatus(
-                        dbUser, accessModuleNameMapper.moduleFromBadge(badgeName))
-                    .map(AccessModuleStatus::getCompletionEpochMillis)
-                    .map(Instant::ofEpochMilli)
-                    .orElse(Instant.EPOCH);
-
-            if (badgeTime.isAfter(dbCompletionTime)) {
-              // First-time badge or renewal: our system recognizes the user as having
-              // completed training right now, though the badge has been issued some
-              // time in the past.
-              return Optional.of(clockNow());
-            }
-
-            // No change
-            return Optional.of(Timestamp.from(dbCompletionTime));
-          };
-
-      Map<DbAccessModuleName, Optional<Timestamp>> completionTimes =
-          Arrays.stream(BadgeName.values())
-              .collect(
-                  Collectors.toMap(
-                      accessModuleNameMapper::moduleFromBadge, determineCompletionTime));
-
-      completionTimes.forEach(
-          (accessModuleName, timestamp) ->
-              accessModuleService.updateCompletionTime(
-                  dbUser, accessModuleName, timestamp.orElse(null)));
-
-      return accessSyncService.updateUserAccessTiers(dbUser, agent);
-    } catch (NumberFormatException e) {
-      log.severe("Incorrect date expire format from Moodle");
-      throw e;
-    } catch (org.pmiops.workbench.moodle.ApiException ex) {
-      if (ex.getCode() == HttpStatus.NOT_FOUND.value()) {
-        log.severe(
-            String.format(
-                "Error while querying Moodle for badges for %s: %s ",
-                dbUser.getUsername(), ex.getMessage()));
-        throw new NotFoundException(ex.getMessage());
-      } else {
-        log.severe(String.format("Error while syncing compliance training: %s", ex.getMessage()));
-      }
-      throw ex;
-    }
-  }
-
   /**
    * Updates the given user's eraCommons-related fields with the NihStatus object returned from FC.
    *
@@ -654,61 +528,6 @@ public class UserServiceImpl implements UserService, GaugeDataCollector {
     DbUser user = userProvider.get();
     FirecloudNihStatus nihStatus = fireCloudService.getNihStatus();
     return setEraCommonsStatus(user, nihStatus, Agent.asUser(user));
-  }
-
-  @Override
-  public void syncTwoFactorAuthStatus() {
-    DbUser user = userProvider.get();
-    syncTwoFactorAuthStatus(user, Agent.asUser(user));
-  }
-
-  @Override
-  public DbUser syncTwoFactorAuthStatus(DbUser targetUser, Agent agent) {
-    return syncTwoFactorAuthStatus(
-        targetUser,
-        agent,
-        directoryService.getUserOrThrow(targetUser.getUsername()).getIsEnrolledIn2Sv());
-  }
-
-  @Override
-  public DbUser syncTwoFactorAuthStatus(DbUser targetUser, Agent agent, boolean isEnrolledIn2FA) {
-    if (isServiceAccount(targetUser)) {
-      // Skip sync for service account user rows.
-      return targetUser;
-    }
-
-    if (isEnrolledIn2FA) {
-      final boolean needsDbCompletionUpdate =
-          accessModuleService
-              .getAccessModuleStatus(targetUser, DbAccessModuleName.TWO_FACTOR_AUTH)
-              .map(status -> status.getCompletionEpochMillis() == null)
-              .orElse(true);
-
-      if (needsDbCompletionUpdate) {
-        accessModuleService.updateCompletionTime(
-            targetUser, DbAccessModuleName.TWO_FACTOR_AUTH, clockNow());
-      }
-    } else {
-      accessModuleService.updateCompletionTime(
-          targetUser, DbAccessModuleName.TWO_FACTOR_AUTH, null);
-    }
-
-    return accessSyncService.updateUserAccessTiers(targetUser, agent);
-  }
-
-  @Override
-  public DbUser syncDuccVersionStatus(DbUser targetUser, Agent agent) {
-    if (isServiceAccount(targetUser)) {
-      // Skip sync for service account user rows.
-      return targetUser;
-    }
-
-    if (!accessModuleService.hasUserSignedACurrentDucc(targetUser)) {
-      accessModuleService.updateCompletionTime(
-          targetUser, DbAccessModuleName.DATA_USER_CODE_OF_CONDUCT, null);
-    }
-
-    return accessSyncService.updateUserAccessTiers(targetUser, agent);
   }
 
   @Override
