+++ api/src/test/java/org/pmiops/workbench/db/dao/UserServiceTest.java
@@ -6,21 +6,14 @@ import static com.google.common.truth.Truth8.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 import static org.pmiops.workbench.access.AccessTierService.REGISTERED_TIER_SHORT_NAME;
 
-import com.google.api.services.directory.model.User;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Random;
 import java.util.function.Supplier;
@@ -33,19 +26,15 @@ import org.pmiops.workbench.access.AccessModuleService;
 import org.pmiops.workbench.access.AccessModuleServiceImpl;
 import org.pmiops.workbench.access.AccessTierServiceImpl;
 import org.pmiops.workbench.access.UserAccessModuleMapperImpl;
-import org.pmiops.workbench.actionaudit.Agent;
 import org.pmiops.workbench.actionaudit.auditors.UserServiceAuditor;
 import org.pmiops.workbench.compliance.ComplianceService;
-import org.pmiops.workbench.compliance.ComplianceService.BadgeName;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.db.model.DbAccessModule;
 import org.pmiops.workbench.db.model.DbAccessModule.DbAccessModuleName;
 import org.pmiops.workbench.db.model.DbAccessTier;
 import org.pmiops.workbench.db.model.DbUser;
-import org.pmiops.workbench.db.model.DbUserCodeOfConductAgreement;
 import org.pmiops.workbench.db.model.DbUserTermsOfService;
 import org.pmiops.workbench.exceptions.BadRequestException;
-import org.pmiops.workbench.exceptions.NotFoundException;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.firecloud.model.FirecloudNihStatus;
 import org.pmiops.workbench.google.DirectoryService;
@@ -53,8 +42,6 @@ import org.pmiops.workbench.institution.InstitutionService;
 import org.pmiops.workbench.mail.MailService;
 import org.pmiops.workbench.model.Authority;
 import org.pmiops.workbench.model.Institution;
-import org.pmiops.workbench.moodle.ApiException;
-import org.pmiops.workbench.moodle.model.BadgeDetailsV2;
 import org.pmiops.workbench.test.FakeClock;
 import org.pmiops.workbench.testconfig.UserServiceTestConfiguration;
 import org.pmiops.workbench.utils.TestMockFactory;
@@ -68,7 +55,6 @@ import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Scope;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
 
 @DataJpaTest
@@ -85,8 +71,6 @@ public class UserServiceTest {
   private static DbAccessTier registeredTier;
   private static List<DbAccessModule> accessModules;
 
-  @MockBean private ComplianceService mockComplianceService;
-  @MockBean private DirectoryService mockDirectoryService;
   @MockBean private FireCloudService mockFireCloudService;
   @MockBean private InstitutionService mockInstitutionService;
   @MockBean private UserServiceAuditor mockUserServiceAuditAdapter;
@@ -103,15 +87,17 @@ public class UserServiceTest {
   @SpyBean private AccessModuleService accessModuleService;
 
   @Import({
+    UserServiceTestConfiguration.class,
+    CommonMappers.class,
     FakeClockConfiguration.class,
     FakeJpaDateTimeConfiguration.class,
-    UserServiceTestConfiguration.class,
-    AccessTierServiceImpl.class,
     AccessModuleServiceImpl.class,
-    CommonMappers.class,
+    AccessTierServiceImpl.class,
     UserAccessModuleMapperImpl.class,
   })
   @MockBean({
+    ComplianceService.class,
+    DirectoryService.class,
     MailService.class,
   })
   @TestConfiguration
@@ -168,144 +154,8 @@ public class UserServiceTest {
         .thenReturn(true);
   }
 
-  @Test
-  public void testSyncComplianceTrainingStatusV2() throws Exception {
-    long issued = fakeClock.instant().getEpochSecond() - 100;
-
-    Map<BadgeName, BadgeDetailsV2> userBadgesByName = new HashMap<>();
-    userBadgesByName.put(
-        BadgeName.REGISTERED_TIER_TRAINING, new BadgeDetailsV2().lastissued(issued).valid(true));
-
-    when(mockComplianceService.getUserBadgesByBadgeName(USERNAME)).thenReturn(userBadgesByName);
-
-    userService.syncComplianceTrainingStatusV2();
-
-    // The user should be updated in the database with a non-empty completion.
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-
-    // Completion timestamp should not change when the method is called again.
-    tick();
-    userService.syncComplianceTrainingStatusV2();
-
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-  }
-
-  @Test
-  public void testUpdateComplianceTrainingStatusV2() throws Exception {
-    long issued = fakeClock.instant().getEpochSecond() - 10;
-    BadgeDetailsV2 retBadge = new BadgeDetailsV2().lastissued(issued).valid(true);
-
-    Map<BadgeName, BadgeDetailsV2> userBadgesByName = new HashMap<>();
-    userBadgesByName.put(BadgeName.REGISTERED_TIER_TRAINING, retBadge);
-
-    when(mockComplianceService.getUserBadgesByBadgeName(USERNAME)).thenReturn(userBadgesByName);
-
-    userService.syncComplianceTrainingStatusV2();
-
-    // The user should be updated in the database with a non-empty completion time.
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-
-    // Deprecate the old training.
-    retBadge.setValid(false);
-
-    // Completion timestamp should be wiped out by the expiry timestamp passing.
-    userService.syncComplianceTrainingStatusV2();
-    assertModuleCompletionEqual(DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, null);
-
-    // The user does a new training.
-    retBadge.lastissued(issued + 5).valid(true);
-
-    // Completion and expiry timestamp should be updated.
-    userService.syncComplianceTrainingStatusV2();
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-
-    // Time passes, user renews training
-    retBadge.lastissued(fakeClock.instant().getEpochSecond() + 1);
-    fakeClock.increment(5000);
-
-    // Completion should be updated to the current time.
-    userService.syncComplianceTrainingStatusV2();
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(fakeClock.instant()));
-  }
-
-  @Test
-  public void testUpdateComplianceTrainingStatusV2_controlled() throws Exception {
-    long issued = fakeClock.instant().getEpochSecond() - 10;
-    BadgeDetailsV2 ctBadge = new BadgeDetailsV2().lastissued(issued).valid(true);
-    Map<BadgeName, BadgeDetailsV2> userBadgesByName =
-        ImmutableMap.<BadgeName, BadgeDetailsV2>builder()
-            .put(
-                BadgeName.REGISTERED_TIER_TRAINING,
-                new BadgeDetailsV2().lastissued(issued).valid(true))
-            .put(BadgeName.CONTROLLED_TIER_TRAINING, ctBadge)
-            .build();
-
-    when(mockComplianceService.getUserBadgesByBadgeName(USERNAME)).thenReturn(userBadgesByName);
-
-    userService.syncComplianceTrainingStatusV2();
-
-    // The user should be updated in the database with a non-empty completion time.
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-    assertModuleCompletionEqual(
-        DbAccessModuleName.CT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-
-    ctBadge.lastissued(fakeClock.instant().getEpochSecond() + 1);
-    fakeClock.increment(5000);
-
-    // Renewing training updates completion.
-    userService.syncComplianceTrainingStatusV2();
-    assertModuleCompletionEqual(
-        DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, Timestamp.from(START_INSTANT));
-    assertModuleCompletionEqual(
-        DbAccessModuleName.CT_COMPLIANCE_TRAINING, user, Timestamp.from(fakeClock.instant()));
-  }
-
   private void tick() {
     fakeClock.increment(CLOCK_INCREMENT_MILLIS);
-  }
-
-  @Test
-  public void testSyncComplianceTrainingStatusNullBadgeV2() throws ApiException {
-    // When Moodle returns an empty RET badge response, we should clear the completion time.
-
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    accessModuleService.updateCompletionTime(
-        user, DbAccessModuleName.RT_COMPLIANCE_TRAINING, new Timestamp(12345));
-
-    // An empty map should be returned when we have no badge information.
-    Map<BadgeName, BadgeDetailsV2> userBadgesByName = new HashMap<>();
-
-    when(mockComplianceService.getUserBadgesByBadgeName(USERNAME)).thenReturn(userBadgesByName);
-
-    userService.syncComplianceTrainingStatusV2();
-    user = userDao.findUserByUsername(USERNAME);
-    assertModuleCompletionEqual(DbAccessModuleName.RT_COMPLIANCE_TRAINING, user, null);
-  }
-
-  @Test
-  public void testSyncComplianceTrainingStatusBadgeNotFoundV2() throws ApiException {
-    // We should propagate a NOT_FOUND exception from the compliance service.
-    when(mockComplianceService.getUserBadgesByBadgeName(USERNAME))
-        .thenThrow(
-            new org.pmiops.workbench.moodle.ApiException(
-                HttpStatus.NOT_FOUND.value(), "user not found"));
-    assertThrows(NotFoundException.class, () -> userService.syncComplianceTrainingStatusV2());
-  }
-
-  @Test
-  public void testSyncComplianceTraining_SkippedForServiceAccountV2() throws ApiException {
-    providedWorkbenchConfig.auth.serviceAccountApiUsers.add(USERNAME);
-    userService.syncComplianceTrainingStatusV2();
-    verifyZeroInteractions(mockComplianceService);
   }
 
   @Test
@@ -391,33 +241,6 @@ public class UserServiceTest {
   }
 
   @Test
-  public void testSyncTwoFactorAuthStatus() {
-    User googleUser = new User();
-    googleUser.setPrimaryEmail(USERNAME);
-    googleUser.setIsEnrolledIn2Sv(true);
-
-    when(mockDirectoryService.getUserOrThrow(USERNAME)).thenReturn(googleUser);
-    userService.syncTwoFactorAuthStatus();
-    // twoFactorAuthCompletionTime should now be set
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    Timestamp twoFactorAuthCompletionTime =
-        getModuleCompletionTime(DbAccessModuleName.TWO_FACTOR_AUTH, user);
-    assertThat(twoFactorAuthCompletionTime).isNotNull();
-
-    // twoFactorAuthCompletionTime should not change when already set
-    tick();
-    userService.syncTwoFactorAuthStatus();
-    assertModuleCompletionEqual(
-        DbAccessModuleName.TWO_FACTOR_AUTH, providedDbUser, twoFactorAuthCompletionTime);
-
-    // unset 2FA in google and check that twoFactorAuthCompletionTime is set to null
-    googleUser.setIsEnrolledIn2Sv(false);
-    userService.syncTwoFactorAuthStatus();
-    user = userDao.findUserByUsername(USERNAME);
-    assertModuleCompletionEqual(DbAccessModuleName.TWO_FACTOR_AUTH, providedDbUser, null);
-  }
-
-  @Test
   public void testSubmitAouTermsOfService() {
     int latestVersion = providedWorkbenchConfig.termsOfService.latestAouVersion;
 
@@ -456,44 +279,6 @@ public class UserServiceTest {
         userTermsOfServiceDao.findFirstByUserIdOrderByTosVersionDesc(user.getUserId());
     assertThat(tosMaybe).isPresent();
     assertThat(tosMaybe.get().getTerraAgreementTime()).isNotNull();
-  }
-
-  @Test
-  public void testSyncDuccVersionStatus_correctVersions() {
-    providedWorkbenchConfig.access.currentDuccVersions = ImmutableList.of(3, 4, 5);
-
-    providedWorkbenchConfig.access.currentDuccVersions.forEach(
-        version -> {
-          DbUser user = userDao.findUserByUsername(USERNAME);
-          user.setDuccAgreement(signDucc(user, version));
-          user = userDao.save(user);
-          userService.syncDuccVersionStatus(user, Agent.asSystem());
-          verify(accessModuleService, never()).updateCompletionTime(any(), any(), any());
-        });
-  }
-
-  @Test
-  public void testSyncDuccVersionStatus_incorrectVersion() {
-    providedWorkbenchConfig.access.currentDuccVersions = ImmutableList.of(3, 4, 5);
-
-    DbUser user = userDao.findUserByUsername(USERNAME);
-    user.setDuccAgreement(signDucc(user, 2));
-    user = userDao.save(user);
-
-    userService.syncDuccVersionStatus(user, Agent.asSystem());
-
-    verify(accessModuleService)
-        .updateCompletionTime(user, DbAccessModuleName.DATA_USER_CODE_OF_CONDUCT, null);
-  }
-
-  @Test
-  public void testSyncDuccVersionStatus_missing() {
-    final DbUser user = userDao.findUserByUsername(USERNAME);
-
-    userService.syncDuccVersionStatus(user, Agent.asSystem());
-
-    verify(accessModuleService)
-        .updateCompletionTime(user, DbAccessModuleName.DATA_USER_CODE_OF_CONDUCT, null);
   }
 
   @Test
@@ -733,10 +518,6 @@ public class UserServiceTest {
     userAccessModuleDao
         .getByUserAndAccessModule(user, accessModuleDao.findOneByName(moduleName).get())
         .ifPresent(module -> assertThat(module.getCompletionTime()).isNull());
-  }
-
-  private DbUserCodeOfConductAgreement signDucc(DbUser dbUser, int version) {
-    return TestMockFactory.createDuccAgreement(dbUser, version, FakeClockConfiguration.NOW);
   }
 
   private DbUser createUserWithAoUTOSVersion(int tosVersion) {
